Refactor to use baldrick
