--- changebot/tests/__init__.py
@@ -1,4 +1,0 @@
-# The tests are divided into two parts - the part that tests the GitHub
-# API interaction, and the part that assumes that the GitHub API classes work
-# and operate based on that. The latter part uses mock GitHub API classes to
-# test the local logic without being dependent on the details of the GitHub API.
