--- changebot/tests/conftest.py
@@ -1,11 +1,0 @@
-import os
-
-os.environ['GITHUB_APP_INTEGRATION_ID'] = '1234'
-os.environ['GITHUB_APP_PRIVATE_KEY'] = 'ABCD'
-os.environ['CRON_TOKEN'] = 'XYZ'
-os.environ['STALE_ISSUE_CLOSE'] = 'TRUE'
-os.environ['STALE_ISSUE_CLOSE_SECONDS'] = '120'
-os.environ['STALE_ISSUE_WARN_SECONDS'] = '60'
-os.environ['STALE_PULL_REQUEST_CLOSE'] = 'TRUE'
-os.environ['STALE_PULL_REQUEST_CLOSE_SECONDS'] = '120'
-os.environ['STALE_PULL_REQUEST_WARN_SECONDS'] = '60'
